pyproject.toml
[project]
name = "clinic-medicine-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.3.1",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
]
